fix(android/api): resolve file upload, dm conversation and persistent unread bugs, and disable realtime service notification

+++ apps/android/app/src/main/java/com/scrymechat/android/ui/share/ShareViewModel.kt
@@ -122,7 +122,7 @@ class ShareViewModel @Inject constructor(
                                 name = uploadResponse.name,
                                 type = uploadResponse.type,
                                 url = uploadResponse.url,
-                                size = uploadResponse.size.toInt()
+                                size = uploadResponse.size
                             )
                         )
                     } else {
